requirements.txt
streamlit==0.78.0
numpy==1.19.5
tensorflow-cpu==2.4.1
Pillow==8.2.0
keras==2.4.3
